requirements.txt
fastapi>=0.111
uvicorn>=0.30
pydantic>=2.7
pandas>=2.2
pyarrow>=16.1
tqdm>=4.66
tenacity>=8.4
google-auth>=2.32
google-cloud-storage>=2.16
gcsfs>=2024.6.1
google-cloud-logging>=3.10
google-cloud-secret-manager>=2.20
google-cloud-aiplatform>=1.66
pytest>=8.2
google-cloud-aiplatform>=1.66.0
sentence-transformers>=3.0
google-cloud-aiplatform>=1.66.0
fastapi
uvicorn[standard]
pydantic
google-cloud-bigquery
google-cloud-aiplatform
gcsfs
tqdm
tenacity
